Move lfp.py to externals, add times variable to example
